Add user setting to show an em-dash as two hyphens
In TUI mode, the typographic styling of proportional fonts is lost. To
alleviate that, in the era of monospace fonts, people used two hyphens
instead of an em dash; they used two spaces after a period. Although the
latter is all but forgotten now and would only annoy a reader from 21st
century, the practice of representing an em dash with two hyphens in
early typewriters is a useful way to distinguish between hyphens and
dashes in monospace fonts. The reader now has the option to choose how
they want to view the em dash.